Refactor how SchoolInfoManager JS is included
